Merged PR 57: Set Mines API URL from environment variables
The API currently routes traffic to the Mines microservice using a harcoded string. This URL needs to be able to be set within K8S to direct traffic to the correct service.

The solution is to switch it to using an environment variable that can be set in the API's K8S deployment yaml

Related work items: #101